refactor(folder): DDD 방법론에 따른 폴더 위치 변경
너무 대량의 폴더도 아니고, 관심사의 분리가 잘 되어있는 폴더구조로 변경
