--- app/src/main/java/com/growit/app/user/domain/User.java
@@ -1,6 +1,0 @@
-package com.growit.app.user.domain;
-
-public class User {
-  private String id;
-  private String name;
-}
