+++ app/src/main/java/com/growit/app/user/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.growit.app.user.domain;
+package com.growit.app.user.domain.user;
 
 public interface UserRepository {
   User findById(String id);
